// packages/contracts/generated/api.ts
export type components = {
    schemas: {
// ...
        Problem: {
            code: string;
            detail: string;
            instance: string;
            /** Format: uuid */
            request_id: string;
            retryable: boolean;
            status: number;
            title: string;
            /** Format: uri */
            type: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
};